feat: add role authorization and more modules (appointment, doctor, doctor_schedule, patient, service, user)

+++ src/specialization/specialization.controller.ts
@@ -1,23 +1,35 @@
-import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
+import { Body, Controller, Delete, ForbiddenException, Get, Param, Post, Req, UseGuards } from '@nestjs/common';
 import { SpecializationService } from './specialization.service';
 import { CreateSpecializationDto } from './dtos/createSpecialization.dto';
+import { AuthGuard } from 'src/common/guards/auth/auth.guard';
+import { RoleName, RoleService } from 'src/roles.service';
+import { JwtPayload } from 'src/common/interfaces/jwt-payload.interface';
 
 @Controller('specialization')
 export class SpecializationController {
-    constructor(private readonly specializationService: SpecializationService){}
+    constructor(
+        private readonly specializationService: SpecializationService,
+        private readonly roleService: RoleService
+    ) { }
 
     @Get()
-    findSpecializations(){
+    findSpecializations() {
         return this.specializationService.findAll()
     }
 
     @Post()
-    createSpecialization(@Body() payload: CreateSpecializationDto){
+    @UseGuards(AuthGuard)
+    createSpecialization(@Body() payload: CreateSpecializationDto, @Req() req: Request) {
+        const { role_id } = req["user"] as JwtPayload
+        if (role_id !== this.roleService.getRoleId(RoleName.ADMIN)) throw new ForbiddenException('You don\'t have enough permission')
         return this.specializationService.create(payload)
     }
 
     @Delete(':id')
-    deleteSpecialization(@Param('id') id: string){
+    @UseGuards(AuthGuard)
+    deleteSpecialization(@Param('id') id: string, @Req() req: Request) {
+        const { role_id } = req["user"] as JwtPayload
+        if (role_id !== this.roleService.getRoleId(RoleName.ADMIN)) throw new ForbiddenException('You don\'t have enough permission')
         return this.specializationService.delete(id)
     }
 }
